requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.5
flask_cors==3.0.10
flask_marshmallow==0.15.0
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
python-dotenv==1.0.1
werkzeug>=3.0.1 # not directly required, pinned by Snyk to avoid a vulnerability
requests == 2.31.0
proofdock-chaos-middleware-flask==0.1.3 #hace que espere 10 seg para simular
proofdock-chaos-middleware-python==0.2.3 # hace que espere 10 seg para simular
psycopg2-binary==2.9.7
saga_py == 1.1.0
tenacity==8.2.3 #se usa para reintentar
Flask-Caching==2.1.0 #cache de objetos
redis==5.0.3
Mako==1.3.5
alembic==1.13.1
flask-migrate==4.0.7